+++ enterprise/api/v1/urls.py
@@ -111,6 +111,11 @@ router.register(
 
 urlpatterns = [
     re_path(
+        r'^enterprise-customer/(?P<enterprise_customer_uuid>[A-Za-z0-9-]+)/admins/(?P<id>[^/.]+)/?$',
+        EnterpriseCustomerAdminViewSet.as_view({'delete': 'delete_admin'}),
+        name='enterprise-customer-delete-admin',
+    ),
+    re_path(
         r'^enterprise_customer_catalog/',
         enterprise_customer_catalog.EnterpriseCustomerCatalogWriteViewSet.as_view(
             {'patch': 'partial_update', 'post': 'create'},
@@ -248,11 +253,6 @@ urlpatterns = [
         r'^(?P<enterprise_uuid>[A-Za-z0-9-]+)/admins$',
         enterprise_admin_members.EnterpriseAdminMembersViewSet.as_view({'get': 'list'}),
         name='enterprise-admin-members',
-    ),
-    re_path(
-        r'^enterprise-customer/(?P<enterprise_customer_uuid>[A-Za-z0-9-]+)/admins/(?P<admin_pk>[A-Za-z0-9-]+)/?$',
-        EnterpriseCustomerAdminViewSet.as_view({'delete': 'delete_admin'}),
-        name='enterprise-customer-admin-delete'
     ),
 ]